poetry-ai/pyproject.toml
[project]
name = "poetry-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.scripts]
poetry-ai = "poetry_ai:main"

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
